+++ develop/config/eslint/config/items/settings/custom.js
@@ -9,7 +9,10 @@ const items = {
      * You can also use `16.0`, `16.3`, etc, if you want to override the detected value.
      * default to latest and warns if missing
      */
-    version: 'detect',
+
+    // It will default to "detect" in the future
+    version: '999.999.999',
+    // version: 'detect',
   },
 };